% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\docType{data}
\name{barrier}
\alias{barrier}
\title{Sample Data for Barrier Option Pricing}
\format{
A dataset with 12 variables , which are:
\describe{
\item{UNDER}{}
\item{STRIKE}{}
\item{BARRIER}{}
\item{EXPIRY}{}
\item{VOL}{}
\item{DY}{}
\item{INTEREST}{}
\item{COR_FX}{}
\item{VOL_FX}{}
\item{FIXING}{}
\item{PriceRCB}{}
\item{PriceQLIB}{}
}
}
\usage{
barrier
}
\description{
Data set provided by Olga containign the caraterisitccs of the Barrier options and the QuantLib and RCB prices of those underlying options.
}
\keyword{datasets}
