% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelBarrier.R
\name{parallelBarrier}
\alias{parallelBarrier}
\title{Price Multiple Barrier Options uUsing Monte Carlo Method and the Parallel Method}
\usage{
parallelBarrier(data = barrier, Option, m, numsim, InOut, nbr.cores = 3L)
}
\arguments{
\item{data}{dataset containing the underlying parameters for the options you want to price. The default set is \code{barrier}.}

\item{Option}{"Put" or "Call"}

\item{m}{nbr of price change simulated}

\item{numsim}{nbr of paths simulated}

\item{InOut}{"In" if you want the price of a Down-In option or "Out" if you want the price od a Down-Out option.}

\item{nbr.cores}{number of cores you want to allow the function to use. It's set at 3 cores (\code{3L}) per default, it can be changed.}
}
\value{
the output is a vector of the expected prices of the underlying options.
}
\description{
\code{parallelBarrier} is a function enhancing the pricing procedure of a Down (Knock-out) Barrier Option (Call or Put has to be specified in the arguments) evaluation using the Monte Carlo method. This function does not work in windows as the function \code{mcmapply} from the \code{parallel} package is not supported by windows.
}
