% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putdown.R
\name{putdown}
\alias{putdown}
\title{Price Put Down Barrier Option using Monte Carlo method}
\usage{
putdown(s, k, v, r, tt, d, B, m, numsim, InOut = "In")
}
\arguments{
\item{s}{= underlying}

\item{k}{= strike}

\item{v}{= volatility}

\item{r}{= riskFreeRate}

\item{tt}{= maturity}

\item{d}{= dividendYield}

\item{B}{= barrier}

\item{m}{= nbr of price change simulated}

\item{numsim}{= nbr of paths simulated}

\item{InOut}{= "In" if you want the price of a Down-In option or "Out" if you want the price od a Down-Out option.}
}
\value{
the output is the expected price for the underlying option.
}
\description{
\code{putdown} is a fuction enhancing the pricinge precedure of a Put Down (Knock-out) Barrier Option evaluation using Monte Carlo method.
}
