% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPrice.R
\name{SPrice}
\alias{SPrice}
\title{Simulate multiple stock prices}
\usage{
SPrice(s, k, v, r, tt, d, m, numsim)
}
\arguments{
\item{s}{underlying}

\item{k}{strike}

\item{v}{volatility}

\item{r}{riskFreeRate}

\item{tt}{maturity}

\item{d}{dividendYield}

\item{m}{nbr of price change simulated}

\item{numsim}{nbr of paths simulated}
}
\value{
the output is a matrix where each row is a different stock (1:numsim), and each column is another time point (1:m).
}
\description{
\code{SPrice} is a function to simulate multiple stock prices with specific characteristics.
}
